refactor: 쿠폰 서비스 예외처리

+++ src/main/java/org/server/remoteclass/service/coupon/CouponServiceImpl.java
@@ -7,6 +7,7 @@ import org.server.remoteclass.dto.coupon.ResponseCouponDto;
 import org.server.remoteclass.entity.Coupon;
 import org.server.remoteclass.entity.FixDiscountCoupon;
 import org.server.remoteclass.entity.RateDiscountCoupon;
+import org.server.remoteclass.exception.ErrorCode;
 import org.server.remoteclass.exception.IdNotExistException;
 import org.server.remoteclass.jpa.CouponRepository;
 import org.server.remoteclass.util.BeanConfiguration;
@@ -32,16 +33,18 @@ public class CouponServiceImpl implements CouponService {
 
     @Override
     public ResponseCouponDto getCouponByCouponId(Long couponId) {
-        if (couponRepository.findByCouponId(couponId).orElse(null) instanceof FixDiscountCoupon) {
+
+        Coupon coupon = (Coupon) couponRepository.findByCouponId(couponId).orElse(null);
+
+        if (coupon instanceof FixDiscountCoupon) {
             log.info("FixDiscount 타입입니다.");
-            return ResponseCouponDto.from((FixDiscountCoupon) couponRepository.findByCouponId(couponId).orElse(null));
+            return ResponseCouponDto.from((FixDiscountCoupon) coupon);
         }
 
         log.info("RateDiscount 타입입니다.");
-        return ResponseCouponDto.from((RateDiscountCoupon) couponRepository.findByCouponId(couponId).orElse(null));
+        return ResponseCouponDto.from((RateDiscountCoupon) coupon);
 
     }
-
 
     @Override
     public List<ResponseCouponDto> getAllCoupons() {
@@ -63,28 +66,23 @@ public class CouponServiceImpl implements CouponService {
     @Override
     @Transactional
     public void deactivateCoupon(Long couponId) {
+        //제너릭(?)으로 구현해서 orElseThrow 문법이 작동하지 않는 것 같습니다. 그래서 orElse 사용하여
+        //null 여부로 에러 처리를 해줬습니다.
         Coupon coupon = (Coupon) couponRepository.findByCouponId(couponId).orElse(null);
+        if(coupon == null){
+            throw new IdNotExistException("존재하지 않는 쿠폰입니다.", ErrorCode.ID_NOT_EXIST);
+        }
         coupon.setCouponValid(false);
-//        Coupon coupon = (Coupon) couponRepository.findByCouponId(couponId)
-//                .orElseThrow(() -> new IdNotExistException("존재하지 않는 쿠폰 번호입니다.", ResultCode.ID_NOT_EXIST)
-//        );
-//        coupon.setCouponValid(false);
-        // 스프링 데이터 JPA가 알아서 해 주는 것 같긴 한데, 일단 save로 업데이트 하였습니다.
-//        return CouponDto.from(couponRepository.save(coupon));
     }
 
     @Override
     @Transactional
     public ResponseCouponDto createCoupon(RequestCouponDto requestCouponDto) {
-        // 0 이하 값이 들어온 경우 예외처리 필요. 나중에 통일시키겠습니다.
-        // 또한 mysql에는 Valid의 default값이 true로 잘 설정되는데, h2 테스트 시에만 false로 출력됩니다.
-        // 그래서 일단 서비스 단에서 true로 초기화 하도록 하겠습니다.
         Coupon coupon = modelMapper.map(requestCouponDto, Coupon.class);
         coupon.setCouponCode(UUID.randomUUID().toString());
         coupon.setCouponValid(true);
         coupon.setCouponValidDays(requestCouponDto.getCouponValidDays());
         return ResponseCouponDto.from((Coupon) couponRepository.save(coupon));
-//        return CouponDto.from(couponRepository.save(coupon));
     }
 
     @Override
@@ -96,9 +94,11 @@ public class CouponServiceImpl implements CouponService {
     @Override
     @Transactional
     public void deleteCoupon(Long couponId) {
-//        Coupon coupon = (Coupon) couponRepository
-//                .findByCouponId(couponId)
-//                .orElseThrow(() -> new IdNotExistException("존재하지 않는 쿠폰 번호입니다", ResultCode.ID_NOT_EXIST));
+        //제너릭(?)으로 구현해서 orElseThrow 문법이 작동하지 않는 것 같습니다. 그래서 orElse 사용했습니다.
+        Coupon coupon = (Coupon) couponRepository.findByCouponId(couponId).orElse(null);
+        if(coupon == null){
+            throw new IdNotExistException("존재하지 않는 쿠폰입니다.", ErrorCode.ID_NOT_EXIST);
+        }
         couponRepository.deleteByCouponId(couponId);
     }
